+++ Scripts/UI/DailyChallengeUI.cs
@@ -42,7 +42,7 @@ namespace ClawRPG.Scripts.UI {
             Visible = false; 
         }
         
-        public override void _Process(float delta) {
+        public override void _Process(double delta) {
             if (!_isVisible) return;
             
             UpdateUI();
